requirements.txt
celery==5.4.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-parler==2.3
django-parler-rest==2.2
django-redis==5.4.0
django-rosetta==0.10.0
django-safedelete==1.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-excel==2.4.0
drf-yasg==1.21.7
gunicorn==22.0.0
Jinja2==3.1.4
openpyxl==3.1.4
pillow==10.3.0
psycopg2-binary==2.9.9
pandas==2.2.2
django-pdfkit==0.3.1
redis==5.0.6
requests==2.32.3
python-docx==1.1.2
tenacity==9.0.0
weasyprint
qrcode
clickhouse-connect
xlsxwriter
pyfcm